# packages/libs/sdk/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agents Control Plane API",
        "version": "1.0.0",
        "description": "REST API for managing agents, jobs, triggers, and webhooks."
    },
...
    "paths": {
...
        "/webhooks/{agentId}": {
            "post": {
                "tags": ["Webhooks"],
                "description": "Invoke an agent via webhook",
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "in": "path",
                        "name": "agentId",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "agent_id": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "payload": {},
                                        "result": {},
                                        "error": {
                                            "type": ["null", "string"]
                                        },
                                        "current_step_id": {
                                            "type": ["null", "string"]
                                        },
                                        "context": {
                                            "type": "object",
                                            "additionalProperties": {}
                                        },
                                        "step_retries": {
                                            "type": "number"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "agent_id",
                                        "status",
                                        "error",
                                        "current_step_id",
                                        "context",
                                        "step_retries",
                                        "created_at",
                                        "updated_at"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Default Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "instance": {
                                            "type": "string"
                                        }
                                    },
                                    "required": ["type", "title", "status", "detail"],
                                    "additionalProperties": false
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}